move projections into package root
