+++ gall-peters.js
@@ -1,24 +1,24 @@
 'use strict'
 
-const h = require('../helpers')
+const h = require('./helpers')
 
-const braun = (point, opt) => {
+const gallPeters = (point, opt) => {
 	point = h.check(point)
 	opt = h.options(opt)
 	if(point.wgs){
 		point = h.addMeridian(point, opt.meridian)
 		return {
 			x: h.rad(point.lon)/(2*Math.PI)+0.5,
-			y: (h.tan(opt.latLimit/2)-h.tan(point.lat/2))/(Math.PI)
+			y: (1-h.sin(point.lat))/Math.PI
 		}
 	}
 	else{
 		const result = {
 			lon: h.deg((2*point.x-1)*Math.PI),
-			lat: h.deg(2*Math.atan(h.tan(opt.latLimit/2)-point.y*Math.PI))
+			lat: h.deg(Math.asin(1-point.y*Math.PI))
 		}
 		return h.addMeridian(result, opt.meridian*(-1))
 	}
 }
 
-module.exports = braun
+module.exports = gallPeters
